Safari-609.1.11.1/LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-generate-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test generating an RSA key using RSASSA-PKCS1-v1_5 algorithm.");

jsTestIsAsync = true;

var algorithmKeyGen = {
    name: "RSASSA-PKCS1-v1_5",
    // RsaKeyGenParams
    modulusLength: 2048,
    publicExponent: new Uint8Array([0x01, 0x00, 0x01]),  // Equivalent to 65537
    hash: "sha-1",
};
var nonExtractable = false;

debug("Generating a key pair...");
crypto.subtle.generateKey(algorithmKeyGen, nonExtractable, ["sign", "verify"]).then(function(result) {
    keyPair = result;
    shouldBe("keyPair.toString()", "'[object Object]'");
    shouldBe("keyPair.publicKey.type", "'public'");
    shouldBe("keyPair.publicKey.extractable", "true");
    shouldBe("keyPair.publicKey.algorithm.name", "'RSASSA-PKCS1-v1_5'");
    shouldBe("keyPair.publicKey.algorithm.modulusLength", "2048");
    shouldBe("bytesToHexString(keyPair.publicKey.algorithm.publicExponent)", "'010001'");
    shouldBe("keyPair.publicKey.algorithm.hash.name", "'SHA-1'");
    shouldBe("keyPair.publicKey.usages", "['verify']");
    shouldBe("keyPair.privateKey.type", "'private'");
    shouldBe("keyPair.privateKey.extractable", "false");
    shouldBe("keyPair.privateKey.algorithm.name", "'RSASSA-PKCS1-v1_5'");
    shouldBe("keyPair.privateKey.algorithm.modulusLength", "2048");
    shouldBe("bytesToHexString(keyPair.privateKey.algorithm.publicExponent)", "'010001'");
    shouldBe("keyPair.privateKey.algorithm.hash.name", "'SHA-1'");
    shouldBe("keyPair.privateKey.usages", "['sign']");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
